+++ src/components/member/tabs/campaigns-tab/campaigns/CampaignsInfo.tsx
@@ -1,7 +1,7 @@
 import { ReferralInformation } from 'api/generated/graphql'
 import { CampaignCodeInput } from 'components/member/tabs/campaigns-tab/campaigns/CampaignCodeInput'
 import { CampaignsRedeemedTable } from 'components/member/tabs/campaigns-tab/campaigns/CampaignsRedeemedTable'
-import { TableHeadline } from 'components/member/tabs/campaigns-tab/referrals/ReferralsInfo'
+import { TableHeadline } from 'components/member/tabs/campaigns-tab/styles'
 import { Card, CardsWrapper } from 'hedvig-ui/card'
 import React from 'react'
